Update and rename map_gui_test.py to map_live.py
